+++ packages/observable-client/src/chainHead/streams/get-runtime-creator.ts
@@ -61,6 +61,7 @@ export interface RuntimeContext {
   accountId: Codec<SS58String>
   assetId: number | null
   getMortalityFromTx: Decoder<Mortality>
+  extVersions: Array<number>
 }
 
 export interface Runtime {