+++ packages/observable-client/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- Export the supported extrinsic versions on the runtime-context
+
 ### Fixed
 
 - Error when unable to track a transaction because we haven't fully recovered from a `stop` event